Updated files for Centos compatibility.

+++ _build/haoscar-binaries/debian/usr/lib/python2.6/binaries/config-image.sh
@@ -49,11 +49,21 @@ assert "SECONDARY_HOSTNAME";
 assert "SECONDARY_IP";
 assert "SUBNET";
 
-# Altering the IP address
-sed s/$PRIMARY_IP/$SECONDARY_IP/g /etc/network/interfaces > /tmp/interfaces
-bak "/etc/network/interfaces";
-mv /tmp/interfaces /etc/network/interfaces
-
+# Executes download commands that are slightly different if a redhat distrobution is detected.
+if [ ! -f /etc/redhat-release ]
+then
+	echo "Deb based distrobution detected....."
+	# Altering the IP address
+	sed s/$PRIMARY_IP/$SECONDARY_IP/g /etc/network/interfaces > /tmp/interfaces
+	bak "/etc/network/interfaces";
+	mv /tmp/interfaces /etc/network/interfaces
+else
+	echo "RPM based distrobution detected....."
+	# Altering the IP address
+	sed s/$PRIMARY_IP/$SECONDARY_IP/g /etc/sysconfig/network-scripts/ifcfg-$HA_ETH > /tmp/ifcfg-$HA_ETH
+	bak "/etc/sysconfig/network-scripts/ifcfg-$HA_ETH";
+	mv /tmp/ifcfg-$HA_ETH /etc/sysconfig/network-scripts/ifcfg-$HA_ETH
+fi	
 
 exit 0; # Exit from the chroot shell
